Components: Unify forms - savings
- apply new unify components to savings pages
- remove legacy components
- fix treasury wrapper components
- add new prop (has-select-modal) to prepare-form
