+++ src/views/treasury/treasury-claim-and-burn-wrapper.vue
@@ -7,6 +7,7 @@
     <prepare-form
       v-if="step === 'prepare'"
       :asset="inputAsset"
+      has-select-modal
       :header-description="
         $t('treasury.claimAndBurn.txtClaimAndBurnPageDescription')
       "
